projects/ToyVLA/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
pydantic>=2.0.0
pytest>=7.4.0
httpx>=0.25.0
